Tech Debt Cleanup

See merge request picovoice/orca!44

+++ app/pv_orca_app.c
@@ -98,7 +98,7 @@ int main(int argc, char *argv[]) {
                 output_path = optarg;
                 break;
             case 'r':
-                speech_rate = atof(optarg);
+                speech_rate = (float) atof(optarg);
                 break;
             case 's':
                 random_state = atoi(optarg);
